+++ assets/js/theme/product.js
@@ -13,6 +13,7 @@ export default class Product extends PageManager {
         super(context);
         this.url = window.location.href;
         this.$reviewLink = $('[data-reveal-id="modal-review-form"]');
+        this.$bulkPricingLink = $('[data-reveal-id="modal-bulk-pricing"]');
     }
 
     onReady() {
@@ -50,11 +51,18 @@ export default class Product extends PageManager {
         });
 
         this.productReviewHandler();
+        this.bulkPricingHandler();
     }
 
     productReviewHandler() {
         if (this.url.indexOf('#write_review') !== -1) {
             this.$reviewLink.trigger('click');
+        }
+    }
+
+    bulkPricingHandler() {
+        if (this.url.indexOf('#bulk_pricing') !== -1) {
+            this.$bulkPricingLink.trigger('click');
         }
     }
 }